+++ src/components/list/index.js
@@ -1,0 +1,1 @@
+export { default as ListModal } from './ListModal';
